app setup done
