--- doobarashop/admin.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-from .models import *
-
-# Register your models here.
-admin.site.register(Categories)
-admin.site.register(Product)
-admin.site.register(Image)
-admin.site.register(ImageAlbum)
-
